requirements.txt
Django==1.11.14
django-webpack-loader==0.6.0
djangorestframework==3.8.2
django-filter==2.0.0
oauth2client==4.1.2
dj-database-url==0.5.0
gunicorn==19.9.0
Markdown==2.6.11
psycopg2==2.7.5
whitenoise==3.3.1
google-api-python-client==1.7.4
google-auth==1.5.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0